Confidence interval estimation

This function estimates based on a parametric bootstrapping procedure (see Details) confidence intervals for a set of signal, noise and SNR spectra obtained from an actual proxy record array.

Usage

EstimateCI(
  spectra,
  f.start = NULL,
  f.end = NULL,
  nmc = 10,
  probs = c(0.1, 0.9),
  df.log = NULL,
  ci.df.log = NULL
)

Arguments

spectra

a list with the spectral objects signal, noise, and snr for an investigated proxy record array, obtained from SeparateSignalFromNoise.

f.start

lower end of the frequency range on which the power-law fit is made on the proxy data (see Details); the default NULL uses the lowest frequency of the proxy spectra.

f.end

as f.start for the upper end; the default NULL uses the uppermost frequency of the proxy spectra.

nmc

integer; the number of replications for the confidence interval estimation.

probs

length-2 numeric vector of probabilities with values in [0,1] defining the confidence interval; defaults to the 10-90 % interval.

df.log

width of the Gaussian kernel in logarithmic frequency units to smooth the spectral estimates of the simulated data; some smoothing is usually necessary to avoid physically implausible negative power occasionally occuring for some frequencies upon estimating the common signal spectrum. It is suggested to use the same amount of smoothing as for the actual proxy data, while setting NULL (the default) suppresses smoothing.

ci.df.log

width of the Gaussian smoothing kernel to smooth the estimated confidence intervals, merely for visual purposes; NULL (the default) suppresses smoothing.

Value

the input spectra object, amended by the confidence intervals for the signal, noise and SNR spectra (element lim.1 gives the upper confidence level, element lim.2 the lower level, respectively).

Details

The parametric bootstrapping procedure for the confidence level estimation is implemented as follows. A power-law fit of the form alpha * f^(-beta) is applied to the actual signal and noise spectra, and the resulting power-law coefficients are used to generate surrogate signal and noise series in a simulated array that mimics the actual proxy record array. This simulated array is replicated nmc times and lower and upper quantiles are calculated across the realizations of the signal, noise and SNR surrogates. Subsequently, the quantiles are scaled to the respective mean estimates and applied multiplicatively to the actual proxy estimates of signal, noise, and SNR to yield the confidence intervals.

See also

Author

Thomas Münch