Plot spectral object

Plot a spectral object on a double-logarithmic scale, optionally adding a transparent confidence interval.

Usage

LPlot(
  x,
  type = "l",
  inverse = FALSE,
  conf = TRUE,
  axes = TRUE,
  col = "black",
  alpha = 0.2,
  removeFirst = 0,
  removeLast = 0,
  xlab = ifelse(inverse, "period", "f"),
  ylab = "PSD",
  xlim = NULL,
  ylim = NULL,
  ...
)

Arguments

x

a spectral object.

type

1-character string giving the type of plot desired: default type "l" makes a line plot, while type "n" produces only the plot frame; see also plot.default.

inverse

if TRUE the x-axis is displayed in units of period (inverse frequency), increasing to the left; defaults to FALSE.

conf

if TRUE (the default) add a transparent confidence interval (has no effect if x contains no confidence limits).

axes

if FALSE the plotting of the x and y axes is suppressed; defaults to TRUE.

col

color for the line plot and the confidence interval.

alpha

transparency level (between 0 and 1) for the confidence interval; defaults to 0.2.

removeFirst

omit removeFirst values on the low frequency side.

removeLast

omit removeLast values on the high frequency side.

xlab

character string for labelling the x-axis.

ylab

character string for labelling the y-axis.

xlim

range of x-axis values; if NULL (the default) it is calculated internally and automatically reversed if inverse = TRUE.

ylim

range of y-axis values; if NULL (the default) it is calculated internally.

...

further graphical parameters passed to plot.

See also

Author

Thomas Laepple