Spectral estimates of proxy record array

Calculate all relevant spectral estimates for a given array of n proxy records. The spectral estimates can be smoothed in logarithmic space and are calculated using Thomson’s multitaper method with three windows with linear detrending before analysis.

Usage

ObtainArraySpectra(cores, res = 1, neff = length(cores), df.log = NULL, ...)

Arguments

cores

a list or a data frame of the proxy data from the record array. If a list is supplied, all elements must be numeric vectors of the same length.

res

the sampling (e.g., temporal) resolution of the proxy data; determines the frequency axis of the spectral estimates.

neff

the effective number of records (e.g. to account for an expected spatial correlation of the local noise). Per default, no spatial noise correlation is assumed and neff is set to the number of proxy records (the length of cores).

df.log

width of the Gaussian kernel in logarithmic frequency units to smooth the spectral estimates; NULL (the default) suppresses smoothing.

...

additional parameters which are passed to the spectral estimation function SpecMTM.

Value

A list of the following components:

single:

a list of n spectral objects (?spec.object) with the spectra of each individual proxy record;

mean:

spectral object of the mean spectrum across all individual spectra;

stack:

spectral object of the spectrum of the average proxy record ("stacked record");

with the attribute array.par: a named vector with information on the proxy record array: number of (effective) records (nc = neff), number of observation points per record (nt), and sampling resolution (res = res).

See also

Author

Thomas Münch