Correlation with common signal

This function calculates the theoretical correlation of a "stacked" proxy record with the common signal depending on the number of records in the stack and the time resolution of the records, given estimates of the average proxy signal and noise spectra.

Usage

ObtainStackCorrelation(input, N = 1, f1 = 2, f2 = "max", limits = NULL)

Arguments

input

a list of the spectral objects (?spec.object) signal and noise, usually to be obtained from a call to SeparateSignalFromNoise.

N

integer vector with the number of records in the assumed stack; correlations are then calculated for stacks with record numbers according to each element of N.

f1

index of the signal (and noise) frequency axis to specify the lower integration limit from which to integrate the spectra; per default the lowest frequency of the spectral estimates is omitted.

f2

as f1 for the upper integration limit; defaults to use the maximum frequency of the given spectral estimates.

limits

numeric vector with a frequency range of the integration: this is an alternative way of specifying the integration limits and overrides the setting by f1 and f2. If not NULL it must be a length-2 vector with the lower integration limit as first and the upper integration limit as second element.

Value

a list of two components:

freq:

numeric vector of frequencies corresponding to the upper ends of the cumulative integrations;

correlation:

a n * m matrix where n corresponds to length(N) and m is given by length(freq) providing the correlation values as a function of the number of averaged records and the record resolution (= increasing upper frequency of the integration).

Details

The function is an implementation of Eqs. (6) and (7) in Münch and Laepple (2018). The integral in (6) is approximated by the cumulative sum of the integration arguments from f.int1 to f.int2, where f.int1 = f1 and f.int2 consecutively increases from f1 to f2.

References

Münch, T. and Laepple, T.: What climate signal is contained in decadal- to centennial-scale isotope variations from Antarctic ice cores? Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.

See also

Author

Thomas Münch