Final DML and WAIS spectra

Internal function to produce the final DML and WAIS spectra (signal, noise, SNR) used for Figs. (3) and (4) in Münch and Laepple (2018). The DML spectra are a combination of the results from the DML1 and DML2 data sets.

Usage

PublicationSNR(
  spec,
  data = c("corr.full", "corr.diff.only", "corr.t.unc.only", "raw"),
  dml.knit.f = 0.1,
  df.log = 0.125
)

Arguments

spec

spectral results for the DML and WAIS firn core data as output from WrapSpectralResults.

data

character string naming the version of the spectral results in spec to use; one of "corr.full", "corr.diff.only", "corr.t.unc.only", and "raw"; defaults to "corr.full".

dml.knit.f

frequency at which to combine the spectra from the DML1 and DML2 data sets (defaults to 1/10 yr^(-1)); DML2 spectra are used for lower frequencies than dml.knit.f, DML1 for the higher frequencies.

df.log

width of Gaussian kernel in logarithmic frequency units for additional smoothing for visual purposes; NULL suppresses smoothing.

Value

A list with the components dml and wais, where each is a list containing three spectral objects (`?spec.object`) and one numeric vector:

signal:

the signal spectrum.

noise:

the noise spectrum.

snr:

the frequency-dependent signal-to-noise ratio.

f.cutoff:

a two-element vector with the index and value of the cutoff frequency from constraining the diffusion correction.

References

Münch, T. and Laepple, T.: What climate signal is contained in decadal- to centennial-scale isotope variations from Antarctic ice cores? Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.

Author

Thomas Münch