Diffusion transfer functions

A dataset containing the calculated diffusion transfer functions for the DML1, DML2 and WAIS records from Münch and Laepple (2018) based on numerical simulations. The transfer function describes the average effect firn diffusion has on the mean of the spectra from a number of N spatially distributed oxygen isotope records.

Usage

diffusion.tf

Format

A named list with the 3 elements dml1, dml2, and wais, where each element is a spectral object:

freq:

the frequency axis in units of 1 / yr,

spec:

the transfer function value at each frequency.

Source

The transfer functions were obtained using ?CalculateDiffusionTF for the site-specific diffusion lengths provided by proxysnr:::diffusion.length; see also the respective package vignette("calculate-transfer-functions"). Isotope records were simulated at semi-annual resolution and the transfer functions interpolated in frequency space to annual resolution. Note here that prior to proxysnr v1.0.0, the transer functions supplied to SeparateSignalFromNoise needed to match in frequency axis the spectra of the analysed data, hence the interpolation step. From proxysnr v1.0.0 onwards this is no longer required, as the frequency axis overlap is checked within SeparateSignalFromNoise, with interpolation taking place there, if needed. The example firn diffusion transfer functions here were, however, left as originally produced to maintain consistency with the results published in Münch and Laepple (2018).

References

Münch, T. and Laepple, T.: What climate signal is contained in decadal- to centennial-scale isotope variations from Antarctic ice cores? Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.

See also