Run simulation for the CI estimation

Get the required parameters from the input spectra (number of observations, observation resolution, number of proxy records, and signal and noise powerlaw parameters) and call the Monte Carlo signal and noise simulation runSurrogates.

Usage

runSimulation(spectra, f.start = NULL, f.end = NULL, nmc = 10, df.log = NULL)

Arguments

spectra

a list with the spectral objects signal, noise, and snr for an investigated proxy record array, obtained from SeparateSignalFromNoise.

f.start

lower end of the frequency range on which the power-law fit is made on the proxy data (see Details); the default NULL uses the lowest frequency of the proxy spectra.

f.end

as f.start for the upper end; the default NULL uses the uppermost frequency of the proxy spectra.

nmc

integer; the number of replications for the confidence interval estimation.

df.log

width of the Gaussian kernel in logarithmic frequency units to smooth the spectral estimates of the simulated data; some smoothing is usually necessary to avoid physically implausible negative power occasionally occuring for some frequencies upon estimating the common signal spectrum. It is suggested to use the same amount of smoothing as for the actual proxy data, while setting NULL (the default) suppresses smoothing.

Value

the output of runSurrogates: a list of length nmc where each list element is one signal and noise estimation realization.

See also

Author

Thomas Münch