Simulate a proxy core array

Simulate an array of proxy records with each record being composed of a signal that is common to all the records in the array and an independent noise component. Both components are simulated such that their spectra follow a power-law of the form `alpha * f^(-beta)` where `f` denotes frequency.

Usage

simCoreArray(signal.par, noise.par, nc, nt)

Arguments

signal.par

a named list with elements `alpha` and `beta` supplying the power-law parameters to simulate the common signal.

noise.par

as signal.par the power-law parameters to simulate the independent noise components.

nc

the number of proxy records (`"cores"`) in the array.

nt

the length of each record in the array.

Value

a list with `nc` elements where each is one simulated proxy record of length `nt`.

Author

Thomas Münch